Corrected gcc-plugin output: ids and constants are 64-bit integers

+++ tools/gcc-bridge/compiler/src/main/java/org/renjin/gcc/gimple/GimpleFunction.java
@@ -35,7 +35,7 @@ import java.util.*;
  * Gimple Function Model
  */
 public class GimpleFunction implements GimpleDecl {
-  private int id;
+  private long id;
   private String name;
   private String mangledName;
   private GimpleType returnType;
@@ -52,11 +52,11 @@ public class GimpleFunction implements GimpleDecl {
   }
 
 
-  public int getId() {
+  public long getId() {
     return id;
   }
 
-  public void setId(int id) {
+  public void setId(long id) {
     this.id = id;
   }
 
@@ -113,7 +113,7 @@ public class GimpleFunction implements GimpleDecl {
 
   public GimpleVarDecl addVarDecl(GimpleType type) {
     // find unused id
-    Set<Integer> usedIds = usedIds();
+    Set<Long> usedIds = usedIds();
     int newId = 1000;
     while(usedIds.contains(newId)) {
       newId++;
@@ -127,8 +127,8 @@ public class GimpleFunction implements GimpleDecl {
     return decl;
   }
   
-  private Set<Integer> usedIds() {
-    Set<Integer> set = new HashSet<>();
+  private Set<Long> usedIds() {
+    Set<Long> set = new HashSet<>();
     for (GimpleVarDecl variableDeclaration : variableDeclarations) {
       set.add(variableDeclaration.getId());
     }
